requirements.txt
torch
torchvision
runpod
diffusers
transformers
accelerate
safetensors
opencv-python-headless
numpy
Pillow
